Add files via upload
In order to run the code yourself, open the folder as a project -> then build all -> then right click project0 (the folder) and do Debug Project -> then hit the restart button in the DEBUG WINDOW and then hit continue. 

The only important codes to look at are in include and src. Those .c and .h files are linked to the files in project0 (which is what the MCU runs), so any code changed in include and src change those files in project0. 

Driverlib is something you have to download on your own.